Fixed path comparison on Windows.
Windows filesystem is case insensitive but boost::filesystem::path is
case sensitive. Use boost::filesystem::equivalent to check for equal
path.
This avoids crashes in h5plotserie for external HDF5 links (the drive
character is sometimes upper and sometimes lowercase which lead to
different path even so the file was the same.)