Switched to a polling condition variable.
boost::interprocess::interprocess_condition is not robust when a process
waiting for the condition is crashing: notify_all is dead-locking.
Hence, implement a own ConditionVarible class which is robust but using
polling instead of putting to sleep and wake-up.
This polling delay is not a problem, the delay is 1/25 second, since it anyway
only used for visual feedback.