Add interface for reading and writing complex datasets
The present interface allows for reading and writing complex vectors and matrices from and to H5 files, respectively. The data is stored with the compound type of H5 using a format that can be read by Octave and Matlab. At the moment the interface consists of global functions. An integration into the class hierarchy of the library should be done next.